chore(Divider): adjust ts types
