+++ components/divider/index.tsx
