+++ components/divider/mobile/index.tsx
