+++ components/divider/style.ts
