+++ components/divider/types.ts
